+++ libopendmi/include/opendmi/module/sun.h
@@ -23,9 +23,4 @@ typedef enum dmi_sun_type
     DMI_TYPE_SUN_MEMORY_DEVICE_EX  = 145, ///< Sun memory device extended information
 } dmi_sun_type_t;
 
-/**
- * @brief Sun extension module.
- */
-extern const dmi_module_t dmi_sun_module;
-
 #endif // !OPENDMI_MODULE_SUN_H