+++ libopendmi/src/module/acer.c
@@ -6,9 +6,18 @@
 //
 #include <opendmi/module/acer.h>
 
-const dmi_module_t dmi_acer_module =
+/**
+ * @brief Acer extension module.
+ */
+static dmi_module_t dmi_acer_module =
 {
     .code     = "acer",
-    .name     = "Acer",
+    .name     = "Acer extensions",
     .entities = nullptr
 };
+
+__attribute__((constructor))
+static void dmi_acer_module_register(void)
+{
+    dmi_module_register(&dmi_acer_module);
+}