+++ libopendmi/src/module/apple.c
@@ -6,9 +6,18 @@
 //
 #include <opendmi/module/apple.h>
 
-const dmi_module_t dmi_apple_module =
+/**
+ * @brief Apple extension module.
+ */
+static dmi_module_t dmi_apple_module =
 {
     .code     = "apple",
-    .name     = "Apple DMI extension",
+    .name     = "Apple extensions",
     .entities = nullptr
 };
+
+__attribute__((constructor))
+static void dmi_apple_module_register(void)
+{
+    dmi_module_register(&dmi_apple_module);
+}