+++ libopendmi/src/module/intel.c
@@ -6,9 +6,18 @@
 //
 #include <opendmi/module/intel.h>
 
-const dmi_module_t dmi_intel_module =
+/**
+ * @brief Intel extension module.
+ */
+static dmi_module_t dmi_intel_module =
 {
     .code     = "intel",
-    .name     = "Intel DMI extension",
+    .name     = "Intel extensions",
     .entities = nullptr
 };
+
+__attribute__((constructor))
+static void dmi_intel_module_register(void)
+{
+    dmi_module_register(&dmi_intel_module);
+}